// prisma/generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserStatus = {
  active: 'active',
  inactive: 'inactive'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const WorkflowStatus = {
  draft: 'draft',
  published: 'published',
  archived: 'archived'
} as const

export type WorkflowStatus = (typeof WorkflowStatus)[keyof typeof WorkflowStatus]


export const ExecutionStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed'
} as const

export type ExecutionStatus = (typeof ExecutionStatus)[keyof typeof ExecutionStatus]


export const PaymentType = {
  stripe: 'stripe',
  sepay: 'sepay',
  momo: 'momo'
} as const

export type PaymentType = (typeof PaymentType)[keyof typeof PaymentType]


export const Currency = {
  USD: 'USD',
  VND: 'VND'
} as const

export type Currency = (typeof Currency)[keyof typeof Currency]


export const LogLevel = {
  info: 'info',
  warning: 'warning',
  error: 'error',
  debug: 'debug'
} as const

export type LogLevel = (typeof LogLevel)[keyof typeof LogLevel]